pyproject.toml
[project]
name = "streamlit-watsonx-chatbot"
version = "0.1.0"
description = "Streamlit chatbot using LangChain + IBM watsonx, uv-managed, with QR popup after N turns."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.32.0",
    "langchain>=0.3.27",
    "langchain-core>=0.3.76",
    "langchain-ibm>=0.3.18",
    "ibm-watsonx-ai>=1.1.0",
    "python-dotenv>=1.0.1",
    "qrcode[pil]>=7.4.2",
    "pillow>=10.3.0",
    "pymilvus>=2.4.0", # for future RAG
    "langchain-openai>=0.3.33",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.4.8",
    "pytest>=8.2.0",
]

[server]
runOnSave = true
